transfer/requirements.txt
boto3
pandas
pylint
pymssql
pytest
python-dotenv